Move declarations and their operators
