+++ komapper-core/src/main/kotlin/org/komapper/core/dsl/declaration/declarations.kt
@@ -1,4 +1,11 @@
-package org.komapper.core.dsl.scope
+package org.komapper.core.dsl.declaration
+
+import org.komapper.core.dsl.scope.HavingScope
+import org.komapper.core.dsl.scope.OnScope
+import org.komapper.core.dsl.scope.SetScope
+import org.komapper.core.dsl.scope.ValuesScope
+import org.komapper.core.dsl.scope.WhenScope
+import org.komapper.core.dsl.scope.WhereScope
 
 typealias OnDeclaration<ENTITY> = OnScope<ENTITY>.() -> Unit
 typealias WhereDeclaration = WhereScope.() -> Unit
