+++ komapper-core/src/main/kotlin/org/komapper/core/dsl/operator/onDeclaration.kt
@@ -1,0 +1,10 @@
+package org.komapper.core.dsl.operator
+
+import org.komapper.core.dsl.declaration.OnDeclaration
+
+infix operator fun <E : Any> OnDeclaration<E>.plus(other: OnDeclaration<E>): OnDeclaration<E> {
+    return {
+        this@plus(this)
+        other(this)
+    }
+}
