+++ komapper-core/src/main/kotlin/org/komapper/core/dsl/operator/setDeclaration.kt
@@ -1,0 +1,10 @@
+package org.komapper.core.dsl.operator
+
+import org.komapper.core.dsl.declaration.SetDeclaration
+
+infix operator fun <E : Any> SetDeclaration<E>.plus(other: SetDeclaration<E>): SetDeclaration<E> {
+    return {
+        this@plus(this)
+        other(this)
+    }
+}
