+++ komapper-core/src/main/kotlin/org/komapper/core/dsl/operator/valuesDeclaration.kt
@@ -1,0 +1,10 @@
+package org.komapper.core.dsl.operator
+
+import org.komapper.core.dsl.declaration.ValuesDeclaration
+
+infix operator fun <E : Any> ValuesDeclaration<E>.plus(other: ValuesDeclaration<E>): ValuesDeclaration<E> {
+    return {
+        this@plus(this)
+        other(this)
+    }
+}
