+++ komapper-core/src/main/kotlin/org/komapper/core/dsl/query/SqlDeleteQuery.kt
@@ -1,9 +1,9 @@
 package org.komapper.core.dsl.query
 
 import org.komapper.core.dsl.context.SqlDeleteContext
+import org.komapper.core.dsl.declaration.WhereDeclaration
 import org.komapper.core.dsl.metamodel.EntityMetamodel
 import org.komapper.core.dsl.options.SqlDeleteOptions
-import org.komapper.core.dsl.scope.WhereDeclaration
 import org.komapper.core.dsl.scope.WhereScope
 import org.komapper.core.dsl.visitor.QueryVisitor
