+++ komapper-core/src/main/kotlin/org/komapper/core/dsl/query/SqlInsertQuery.kt
@@ -1,10 +1,10 @@
 package org.komapper.core.dsl.query
 
 import org.komapper.core.dsl.context.SqlInsertContext
+import org.komapper.core.dsl.declaration.ValuesDeclaration
 import org.komapper.core.dsl.element.Values
 import org.komapper.core.dsl.metamodel.EntityMetamodel
 import org.komapper.core.dsl.options.SqlInsertOptions
-import org.komapper.core.dsl.scope.ValuesDeclaration
 import org.komapper.core.dsl.scope.ValuesScope
 import org.komapper.core.dsl.visitor.QueryVisitor
