+++ komapper-core/src/main/kotlin/org/komapper/core/dsl/query/SqlInsertQueryBuilder.kt
@@ -1,7 +1,7 @@
 package org.komapper.core.dsl.query
 
 import org.komapper.core.ThreadSafe
-import org.komapper.core.dsl.scope.ValuesDeclaration
+import org.komapper.core.dsl.declaration.ValuesDeclaration
 
 @ThreadSafe
 interface SqlInsertQueryBuilder<T : Any> {
