+++ komapper-core/src/main/kotlin/org/komapper/core/dsl/query/SqlUpdateQuery.kt
@@ -1,11 +1,11 @@
 package org.komapper.core.dsl.query
 
 import org.komapper.core.dsl.context.SqlUpdateContext
+import org.komapper.core.dsl.declaration.SetDeclaration
+import org.komapper.core.dsl.declaration.WhereDeclaration
 import org.komapper.core.dsl.metamodel.EntityMetamodel
 import org.komapper.core.dsl.options.SqlUpdateOptions
-import org.komapper.core.dsl.scope.SetDeclaration
 import org.komapper.core.dsl.scope.SetScope
-import org.komapper.core.dsl.scope.WhereDeclaration
 import org.komapper.core.dsl.scope.WhereScope
 import org.komapper.core.dsl.visitor.QueryVisitor
