+++ komapper-core/src/main/kotlin/org/komapper/core/dsl/scope/WhenScope.kt
@@ -1,6 +1,7 @@
 package org.komapper.core.dsl.scope
 
 import org.komapper.core.Scope
+import org.komapper.core.dsl.declaration.WhenDeclaration
 import org.komapper.core.dsl.element.Criterion
 
 @Scope
@@ -8,29 +9,6 @@ class WhenScope(
     private val support: FilterScopeSupport<WhenScope> = FilterScopeSupport { WhenScope() }
 ) : FilterScope by support,
     List<Criterion> by support {
-
-    companion object {
-        infix operator fun WhenDeclaration.plus(other: WhenDeclaration): WhenDeclaration {
-            return {
-                this@plus(this)
-                other(this)
-            }
-        }
-
-        infix fun WhenDeclaration.and(other: WhenDeclaration): WhenDeclaration {
-            return {
-                this@and(this)
-                and { other(this) }
-            }
-        }
-
-        infix fun WhenDeclaration.or(other: WhenDeclaration): WhenDeclaration {
-            return {
-                this@or(this)
-                or { other(this) }
-            }
-        }
-    }
 
     fun and(declaration: WhenDeclaration) {
         support.addCriteria(declaration, Criterion::And)
