+++ komapper-core/src/main/kotlin/org/komapper/core/dsl/scope/WhereScope.kt
@@ -1,6 +1,7 @@
 package org.komapper.core.dsl.scope
 
 import org.komapper.core.Scope
+import org.komapper.core.dsl.declaration.WhereDeclaration
 import org.komapper.core.dsl.element.Criterion
 
 @Scope
@@ -8,29 +9,6 @@ class WhereScope(
     private val support: FilterScopeSupport<WhereScope> = FilterScopeSupport { WhereScope() }
 ) : FilterScope by support,
     List<Criterion> by support {
-
-    companion object {
-        infix operator fun WhereDeclaration.plus(other: WhereDeclaration): WhereDeclaration {
-            return {
-                this@plus(this)
-                other(this)
-            }
-        }
-
-        infix fun WhereDeclaration.and(other: WhereDeclaration): WhereDeclaration {
-            return {
-                this@and(this)
-                and { other(this) }
-            }
-        }
-
-        infix fun WhereDeclaration.or(other: WhereDeclaration): WhereDeclaration {
-            return {
-                this@or(this)
-                or { other(this) }
-            }
-        }
-    }
 
     fun and(declaration: WhereDeclaration) {
         support.addCriteria(declaration, Criterion::And)
