+++ komapper-core/src/test/kotlin/org/komapper/core/dsl/scope/WhereScopeTest.kt
@@ -2,12 +2,13 @@ package org.komapper.core.dsl.scope
 
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Test
+import org.komapper.core.dsl.declaration.WhereDeclaration
 import org.komapper.core.dsl.element.Criterion
 import org.komapper.core.dsl.element.Operand
 import org.komapper.core.dsl.metamodel.PropertyMetamodelStub
-import org.komapper.core.dsl.scope.WhereScope.Companion.and
-import org.komapper.core.dsl.scope.WhereScope.Companion.or
-import org.komapper.core.dsl.scope.WhereScope.Companion.plus
+import org.komapper.core.dsl.operator.and
+import org.komapper.core.dsl.operator.or
+import org.komapper.core.dsl.operator.plus
 
 internal class WhereScopeTest {
